fix a small bug, extend the search time to 2 days, and update README.
I shared my progress in standup and found a bug, and cleaned up some parts of the code to make it more readable.

I still need to update the README with instructions on how to make an single binary executable so that it can be
made into a cronjob to run everyday.